:art: refactor: move `DummyLandCoverChangeAnalyzer` into its own module
